Port architecture (#4)
* Add prettier

* Bite the bullet, unify Leaflet

My hands hurt from all the JSON encoding. It is pretty silly that
we cannot leverage the automagical Elm Json conversion for these
simple values, in the way that using one port per function does.
Oh well.

* Add xstream

Debatable whether it is useful atm, but I like having a common
interface for subscriptions going forward

* Move to Typescript :tada:

Some rough edges still (and I need to type things and turn on
strict), but this is cool

* Fix lazy loading

* Add pouchdb-server to devDeps

* Minor cleanup in index.ts

* One click deploy

Can't see why I didn't do this earlier

* Move out Leaflet into module, use streams :tada:

Feels like programming in a weird Elm-Cycle frankenstein's monster
where update is somehow just effects and in-place mutations.
Tbh, it could be done in an immutable fashion easily, but I'm
unaware of the performance implications (or lack thereof). Thus,
since we're in JS, I'm taking a middle ground approach.

* Pass in stream to Leaflet module

I am not sure which implementation is better, but I like the idea
of Leaflet being unaware of where the messages come from.

* Initial port architecture for Pouch

* Listen to DB instead of the connecion for changes

* Remove build artefact

* Bump pouchdb-authentication and pouchdb

Better types and such

* First pass at completing types

* More types wheeeeee

* "All" types done :sweat_smile:

* Finalise types, add card export in the mix

* Re-structure js side (util, export)

* Use 'EphemeralDB' in place of PouchDB.Database<..>

This allows specifying doc types more easily down the line

* Fix export Msg, update remoteDB and sync properly

* Fix naming of export chunk

Additional caches are set to use *-chunk-*